Use Bigint for ID types

+++ src/models.rs
@@ -15,7 +15,7 @@ pub struct NewUser {
 #[diesel(table_name = crate::schema::users)]
 #[diesel(check_for_backend(diesel::sqlite::Sqlite))]
 pub struct User {
-    pub id: i32,
+    pub id: i64,
     pub user_name: String,
     pub password_hash: String,
     //pub created:NaiveDateTime,
@@ -28,7 +28,7 @@ pub struct User {
 pub struct NewSession {
     pub token: String,
     pub token_pass: String,
-    pub user_id: i32,
+    pub user_id: i64,
     pub expires: NaiveDateTime,
 }
 
@@ -36,11 +36,11 @@ pub struct NewSession {
 #[diesel(table_name = crate::schema::files)]
 #[diesel(check_for_backend(diesel::sqlite::Sqlite))]
 pub struct File {
-    pub id: i32,
+    pub id: i64,
     pub file_type: CT::FileType,
     pub file_name: String,
-    pub file_size: i32,
-    pub chunk_size: i32,
+    pub file_size: i64,
+    pub chunk_size: i64,
     pub chunks_loaded: i32,
     pub file_hash: String,
     pub completed: Option<NaiveDateTime>,
